+++ src/components/icon-button/icon-button.styles.ts
@@ -1,5 +1,6 @@
 import { css } from 'lit';
 import componentStyles from '../../styles/component.styles';
+import { focusVisibleSelector } from '../../internal/focus-visible';
 
 export default css`
   ${componentStyles}
@@ -41,7 +42,7 @@ export default css`
     cursor: not-allowed;
   }
 
-  .focus-visible.icon-button:focus {
+  .icon-button${focusVisibleSelector} {
     box-shadow: 0 0 0 var(--sl-focus-ring-width) rgb(var(--sl-color-primary-500) / var(--sl-focus-ring-alpha));
   }
 `;