+++ src/components/rating/rating.styles.ts
@@ -1,5 +1,6 @@
 import { css } from 'lit';
 import componentStyles from '../../styles/component.styles';
+import { focusVisibleSelector } from '../../internal/focus-visible';
 
 export default css`
   ${componentStyles}
@@ -24,7 +25,7 @@ export default css`
     outline: none;
   }
 
-  .rating.focus-visible:focus {
+  .rating${focusVisibleSelector} {
     box-shadow: 0 0 0 var(--sl-focus-ring-width) rgb(var(--sl-color-primary-500) / var(--sl-focus-ring-alpha));
   }