+++ src/internal/focus-visible.ts
@@ -1,53 +1,26 @@
+import { unsafeCSS } from 'lit';
+
 //
-// Simulates :focus-visible behavior on an element by watching for certain keyboard and mouse heuristics and toggling a
-// `focus-visible` class. Works at the component level so no global polyfill is necessary.
+// Determines if the current browser supports :focus-visible
 //
-// This will eventually be removed pending better :focus-visible support: https://caniuse.com/#search=focus-visible
+export const hasFocusVisible = (() => {
+  const style = document.createElement('style');
+  let isSupported;
+
+  try {
+    document.head.appendChild(style);
+    style.sheet!.insertRule(':focus-visible { color: inherit }');
+    isSupported = true;
+  } catch {
+    isSupported = false;
+  } finally {
+    style.remove();
+  }
+
+  return isSupported;
+})();
+
 //
-const listeners = new WeakMap();
-
-interface ObserveOptions {
-  visible: () => void;
-  notVisible: () => void;
-}
-
-export function observe(el: HTMLElement, options?: ObserveOptions) {
-  const keys = ['Tab', 'ArrowUp', 'ArrowDown', 'ArrowLeft', 'ArrowRight', 'Home', 'End', 'PageDown', 'PageUp'];
-  const is = (event: KeyboardEvent) => {
-    if (keys.includes(event.key)) {
-      el.classList.add('focus-visible');
-
-      if (options?.visible) {
-        options.visible();
-      }
-    }
-  };
-  const isNot = () => {
-    el.classList.remove('focus-visible');
-
-    if (options?.notVisible) {
-      options.notVisible();
-    }
-  };
-  listeners.set(el, { is, isNot });
-
-  el.addEventListener('keydown', is);
-  el.addEventListener('keyup', is);
-  el.addEventListener('mousedown', isNot);
-  el.addEventListener('mouseup', isNot);
-}
-
-export function unobserve(el: HTMLElement) {
-  const { is, isNot } = listeners.get(el);
-
-  el.classList.remove('focus-visible');
-  el.removeEventListener('keydown', is);
-  el.removeEventListener('keyup', is);
-  el.removeEventListener('mousedown', isNot);
-  el.removeEventListener('mouseup', isNot);
-}
-
-export const focusVisible = {
-  observe,
-  unobserve
-};
+// A selector for Lit stylesheets that outputs `:focus-visible` if the browser supports it and `:focus` otherwise
+//
+export const focusVisibleSelector = unsafeCSS(hasFocusVisible ? ':focus-visible' : ':focus');